+++ distrib-kv-store/scripts/raft/test-cluster.sh
